bugfix: fix a bug where custom serializer blocks can fail if after a list
This commit fixes a bug where the rendering of a node can fail
if it has fields not supported by Block and follows directly after
a list item.

The list item logic would pass in the first node after the list as context
and cast it to a Block. This fails if the node has fields not supported
by Block. Which is the case for custom serializer blocks.

The context is actually not used, and removing it solves the bug.